+++ api/controllers/PlaylistController.js
@@ -4,6 +4,7 @@ class PlaylistController {
     static async listAll(req, res) {
         try {
             const allPlaylists = await database.Playlists.findAll()
+        
             return res.status(200).json(allPlaylists)
         } catch (error) {
             return res.status(500).json(error.message)
@@ -14,6 +15,7 @@ class PlaylistController {
         const { id } = req.params
         try {
             const onePlaylist = await database.Playlists.findOne( { where: { id: Number(id) } } )
+ 
             return res.status(200).json(onePlaylist)
         } catch (error) {
             return res.status(500).json(error.message)
@@ -24,6 +26,7 @@ class PlaylistController {
         const newPlaylist = req.body
         try {
             const newPlaylistCreated = await database.Playlists.create(newPlaylist)
+           
             return res.status(201).json(newPlaylistCreated)          
         } catch (error) {
             return res.status(500).json(error.message)
@@ -36,6 +39,7 @@ class PlaylistController {
         try {
             await database.Playlists.update(newInfo, { where: { id: Number(id) } })
             const playlistUpdated = await database.Playlists.findOne( { where: { id: Number(id) } } )
+          
             return res.status(200).json(playlistUpdated)
         } catch (error) {
             return res.status(500).json(error.message)
@@ -45,6 +49,14 @@ class PlaylistController {
     static async delete(req, res) {
         const { id } = req.params
         try {
+            const playlistExiste = await database.Playlists.findOne({
+                where: { id: Number(id) }
+            })
+
+            if(!playlistExiste){
+                return res.status(404).json({ mensagem: 'O registro não existe.' })
+            }
+
             await database.Playlists.destroy( { where: { id: Number(id) } } )
             return res.status(200).json( { mensagem: "O registro foi deletado" } )
         } catch (error) {
@@ -55,18 +67,27 @@ class PlaylistController {
     static async listOneMusica(req, res) {
         const { id_playlist, id_musica } = req.params
         try {
-            const oneMusica = await database.Playlists_musicas.findOne({
+            const existeMusica = await database.Playlists_musicas.findOne({
                 where: {
                     id_playlist: Number(id_playlist),
                     id_musica: Number(id_musica)
                 }
             })
-            if(!oneMusica) {
-                return res.status(200).json({ mensagem: 'Não foi encontrado nenhum registro' })
-            } else {
-                const descrMusica = await database.Musicas.findOne({ where: { id: Number(oneMusica.id_musica) } })
-                return res.status(200).json(descrMusica)
+            if(!existeMusica) {
+                return res.status(404).json({ mensagem: 'Não foi encontrado nenhum registro' })
             }
+
+            const infoMusica = await database.Musicas.findOne({ where: { id: Number(id_musica) } })
+            const infoPlaylist = await database.Playlists.findOne({ where: { id: Number(id_playlist) } })
+           
+            const musicaFormatada = {
+                id_playlist: id_playlist,
+                id_musica: id_musica, 
+                playlist: infoPlaylist.nome,
+                musica: infoMusica.nome
+            }
+
+            return res.status(200).json(musicaFormatada)
         } catch (error) {
             return res.status(500).json(error.message)
         }
@@ -84,25 +105,55 @@ class PlaylistController {
                 }
             })
             if(!allMusicas) {
-                return res.status(200).json({ mensagem: 'Não foi encontrado nenhum registro' })
-            } else {
-                const musicasFormatas = allMusicas.map(musica => ({
-                    id_musica: musica.id_musica, 
-                    id_playlist: musica.id_playlist, 
-                    nome: musica.Musica.nome
-                }))
-                return res.status(200).json(musicasFormatas)
-            }
+                return res.status(500).json({ mensagem: 'Não foi encontrado nenhum registro' })
+            } 
+            const musicasFormatas = allMusicas.map(musica => ({
+                id_musica: musica.id_musica, 
+                id_playlist: musica.id_playlist, 
+                nome: musica.Musica.nome
+            }))
+
+            return res.status(200).json(musicasFormatas)
         } catch (error) {
             return res.status(500).json(error.message)
         }
     }
 
     static async insertMusicIntoPlaylist(req, res) {
+        const newRegister = req.body
         try {
-            const newRegister = req.body
+
+            const newMusica = await database.Musicas.findOne({
+                where: {id: Number(newRegister.id_musica)}
+            })
+            
+            if(!newMusica) {
+                return res.status(500).json({ mensagem: 'A música não foi encontrada' })    
+            } 
             const newRegisterCreated = await database.Playlists_musicas.create(newRegister)
+            
             return res.status(200).json(newRegisterCreated)
+        } catch (error) {
+            return res.status(500).json(error.message)
+        }
+    }
+
+    static async deleteMusicInPlaylist(req, res) {
+        const { id_playlist, id_musica } = req.params
+        try { 
+            const musicaExiste = await database.Musicas.findOne({
+                where: { id: Number(id_musica) }
+            })
+
+            if(!musicaExiste){
+                return res.status(404).json({ mensagem: 'O registro não existe.' })
+            }
+
+            await database.Playlists_musicas.destroy( { where: {
+                id_musica: Number(id_musica),
+                id_playlist: Number(id_playlist)
+            }})
+            return res.status(200).json( { mensagem: "O registro foi deletado" } )
         } catch (error) {
             return res.status(500).json(error.message)
         }
